+++ client/src/components/CartItem/CartItem.jsx
@@ -1,16 +1,18 @@
-import React from "react";
-import "./CartItem.scss";
+import React from 'react'
+import './CartItem.scss'
 
 const CartItem = ({ item: { name, imageUrl, price, quantity } }) => {
   return (
-    <div className="cart-item">
+    <div className='cart-item'>
       <img src={imageUrl} alt={name} />
-      <div className="item-details">
-        <span className="name">{name}</span>
-        <span className="price">{quantity} x ${price}</span>
+      <div className='item-details'>
+        <span className='name'>{name}</span>
+        <span className='price'>
+          {quantity} x ${price}
+        </span>
       </div>
     </div>
-  );
-};
+  )
+}
 
-export default CartItem;
+export default React.memo(CartItem)