+++ client/src/components/WithSpinner/Spinner.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { SpinnerContainer, SpinnerOverlay } from './WithSpinner.styles'
+
+export default () => (
+  <SpinnerOverlay>
+    <SpinnerContainer />
+  </SpinnerOverlay>
+)