+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-move or update store pictures data
+- apply lazy loading and chunking using react lazy and suspense
+- added ErrorBoundary for HomePage for a callback component